chore: reuse error codes; new followup analysis in manager class [DHIS2-10658]
