+++ dhis-2/dhis-services/dhis-service-validation/src/main/java/org/hisp/dhis/dataanalysis/jdbc/JdbcDataAnalysisStore.java
@@ -28,11 +28,8 @@
 package org.hisp.dhis.dataanalysis.jdbc;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static java.util.Collections.emptyList;
 import static org.hisp.dhis.common.IdentifiableObjectUtils.getIdentifiers;
-import static org.hisp.dhis.commons.collection.CollectionUtils.isEmpty;
 import static org.hisp.dhis.commons.util.TextUtils.getCommaDelimitedString;
-import static org.hisp.dhis.query.JpaQueryUtils.generateHqlQueryForSharingCheck;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -42,24 +39,17 @@ import java.util.Map;
 
 import lombok.extern.slf4j.Slf4j;
 
-import org.hibernate.SessionFactory;
-import org.hibernate.query.Query;
 import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.commons.collection.PaginatedList;
 import org.hisp.dhis.commons.util.TextUtils;
 import org.hisp.dhis.dataanalysis.DataAnalysisMeasures;
 import org.hisp.dhis.dataanalysis.DataAnalysisStore;
-import org.hisp.dhis.dataanalysis.FollowupAnalysisRequest;
-import org.hisp.dhis.dataanalysis.FollowupValue;
 import org.hisp.dhis.dataelement.DataElement;
 import org.hisp.dhis.datavalue.DeflatedDataValue;
 import org.hisp.dhis.jdbc.StatementBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.period.Period;
-import org.hisp.dhis.period.PeriodType;
-import org.hisp.dhis.security.acl.AclService;
 import org.hisp.dhis.system.objectmapper.DeflatedDataValueNameMinMaxRowMapper;
-import org.hisp.dhis.user.User;
 import org.hisp.dhis.util.DateUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -82,19 +72,14 @@ public class JdbcDataAnalysisStore implements DataAnalysisStore
      */
     private final JdbcTemplate jdbcTemplate;
 
-    private final SessionFactory sessionFactory;
-
     public JdbcDataAnalysisStore( StatementBuilder statementBuilder,
-        @Qualifier( "readOnlyJdbcTemplate" ) JdbcTemplate jdbcTemplate,
-        SessionFactory sessionFactory )
+        @Qualifier( "readOnlyJdbcTemplate" ) JdbcTemplate jdbcTemplate )
     {
         checkNotNull( statementBuilder );
         checkNotNull( jdbcTemplate );
-        checkNotNull( sessionFactory );
 
         this.statementBuilder = statementBuilder;
         this.jdbcTemplate = jdbcTemplate;
-        this.sessionFactory = sessionFactory;
     }
 
     @Override
@@ -301,91 +286,5 @@ public class JdbcDataAnalysisStore implements DataAnalysisStore
         sql += statementBuilder.limitRecord( 0, limit );
 
         return jdbcTemplate.query( sql, new DeflatedDataValueNameMinMaxRowMapper( null, null ) );
-    }
-
-    /**
-     * HQL query used to select {@link FollowupValue}s.
-     */
-    private static final String FOLLOWUP_VALUE_HQL = "select new org.hisp.dhis.dataanalysis.FollowupValue(" +
-        "de.uid, de.name, " +
-        "pt.class, " +
-        "pe.startDate, pe.endDate, " +
-        "ou.uid, ou.name, ou.path, " +
-        "coc.uid, coc.name, " +
-        "aoc.uid, aoc.name, " +
-        "dv.value, dv.storedBy, dv.lastUpdated, dv.created, dv.comment, " +
-        "mm.min, mm.max " +
-        ")" +
-        "from DataValue dv " +
-        "left join MinMaxDataElement mm on (dv.dataElement.id = mm.dataElement.id and dv.categoryOptionCombo.id = mm.optionCombo.id and dv.source.id = mm.source.id) "
-        +
-        "inner join DataElement de on dv.dataElement.id = de.id " +
-        "inner join Period pe on dv.period.id = pe.id " +
-        "inner join PeriodType pt on pe.periodType.id = pt.id " +
-        "inner join OrganisationUnit ou on dv.source.id = ou.id " +
-        "inner join CategoryOptionCombo coc on dv.categoryOptionCombo.id = coc.id " +
-        "inner join CategoryOptionCombo aoc on dv.attributeOptionCombo.id = aoc.id " +
-        "where de.uid in (:de_ids) and (<<sharing>>)" +
-        "and dv.followup = true and dv.deleted is false " +
-        "and coc.uid in (:coc_ids) " +
-        "and pe.startDate >= :startDate and pe.endDate <= :endDate " +
-        "and exists (select 1 from OrganisationUnit parent where parent.uid in (:ou_ids) and ou.path like concat(parent.path, '%'))";
-
-    /**
-     * HQL query potentially used to resolve all
-     * {@link org.hisp.dhis.dataelement.DataElement} UIDS from
-     * {@link org.hisp.dhis.dataset.DataSet} UIDs in context of
-     * {@link #FOLLOWUP_VALUE_HQL} query.
-     */
-    private static final String DATA_ELEMENT_UIDS_BY_DATA_SET_UIDS_HQL = "select dse.dataElement.uid " +
-        "from DataSetElement dse " +
-        "where dse.dataSet.uid in (:ds_ids)" +
-        "and dse.dataElement.valueType in ('INTEGER', 'INTEGER_POSITIVE', 'INTEGER_NEGATIVE', 'INTEGER_ZERO_OR_POSITIVE', 'NUMBER', 'UNIT_INTERVAL', 'PERCENTAGE')";
-
-    private static final String CATEGORY_OPTION_COMBO_UIDS_BY_DATE_ELEMENT_UIDS_HQL = "select coc.uid " +
-        "from CategoryOptionCombo coc " +
-        "where coc.categoryCombo.id in (select de.categoryCombo.id from DataElement de where de.uid in (:de_ids))";
-
-    @Override
-    public List<FollowupValue> getFollowupDataValues( User currentUser, FollowupAnalysisRequest request )
-    {
-        if ( isEmpty( request.getDe() ) && !isEmpty( request.getDs() ) )
-        {
-            request.setDe( sessionFactory.getCurrentSession()
-                .createQuery( DATA_ELEMENT_UIDS_BY_DATA_SET_UIDS_HQL, String.class )
-                .setParameter( "ds_ids", request.getDs() ).list() );
-        }
-        if ( !isEmpty( request.getDe() ) && isEmpty( request.getCoc() ) )
-        {
-            request.setCoc( sessionFactory.getCurrentSession()
-                .createQuery( CATEGORY_OPTION_COMBO_UIDS_BY_DATE_ELEMENT_UIDS_HQL, String.class )
-                .setParameter( "de_ids", request.getDe() ).list() );
-        }
-        if ( isEmpty( request.getDe() ) || isEmpty( request.getCoc() )
-            || isEmpty( request.getOu() ) )
-        {
-            return emptyList();
-        }
-        if ( request.getStartDate() == null && request.getPe() != null )
-        {
-            request.setStartDate( PeriodType.getPeriodFromIsoString( request.getPe() ).getStartDate() );
-        }
-        if ( request.getEndDate() == null && request.getPe() != null )
-        {
-            request.setEndDate( PeriodType.getPeriodFromIsoString( request.getPe() ).getEndDate() );
-        }
-
-        Query<FollowupValue> query = sessionFactory.getCurrentSession().createQuery(
-            FOLLOWUP_VALUE_HQL.replace( "<<sharing>>",
-                generateHqlQueryForSharingCheck( "de", currentUser, AclService.LIKE_READ_METADATA ) ),
-            FollowupValue.class );
-
-        query.setParameter( "ou_ids", request.getOu() );
-        query.setParameter( "de_ids", request.getDe() );
-        query.setParameter( "coc_ids", request.getCoc() );
-        query.setParameter( "startDate", request.getStartDate() );
-        query.setParameter( "endDate", request.getEndDate() );
-        query.setMaxResults( request.getMaxResults() );
-        return query.setCacheable( false ).list();
     }
 }
